test(reef): split receipt flow coverage
